Remove need for inits
Client and Driver are fully usable after calling New()
