+++ constraint/pkg/client/drivers/local/local.go
@@ -6,12 +6,9 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
-	"net/http"
 	"sort"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/open-policy-agent/frameworks/constraint/pkg/handler/handlertest"
 
@@ -70,6 +67,17 @@ func New(args ...Arg) (*Driver, error) {
 
 	d.compiler.WithCapabilities(d.capabilities)
 
+	if d.providerCache != nil {
+		rego.RegisterBuiltin1(
+			&rego.Function{
+				Name:    "external_data",
+				Decl:    opatypes.NewFunction(opatypes.Args(opatypes.A), opatypes.A),
+				Memoize: true,
+			},
+			externalDataBuiltin(d),
+		)
+	}
+
 	return d, nil
 }
 
@@ -77,6 +85,10 @@ var _ drivers.Driver = &Driver{}
 
 type Driver struct {
 	mtx sync.RWMutex
+
+	// compilers is a map from target name to a map from Template Kind to the
+	// compiler for that Template.
+	compilers map[string]map[string]*ast.Compiler
 
 	compiler      *ast.Compiler
 	modules       map[string]*ast.Module
@@ -89,67 +101,6 @@ type Driver struct {
 	externs       []string
 
 	kinds map[string]bool
-}
-
-func (d *Driver) Init() error {
-	d.mtx.Lock()
-	defer d.mtx.Unlock()
-
-	if d.providerCache != nil {
-		rego.RegisterBuiltin1(
-			&rego.Function{
-				Name:    "external_data",
-				Decl:    opatypes.NewFunction(opatypes.Args(opatypes.A), opatypes.A),
-				Memoize: true,
-			},
-			func(bctx rego.BuiltinContext, regorequest *ast.Term) (*ast.Term, error) {
-				var regoReq externaldata.RegoRequest
-				if err := ast.As(regorequest.Value, &regoReq); err != nil {
-					return nil, err
-				}
-
-				provider, err := d.providerCache.Get(regoReq.ProviderName)
-				if err != nil {
-					return externaldata.HandleError(http.StatusBadRequest, err)
-				}
-
-				externaldataRequest := externaldata.NewProviderRequest(regoReq.Keys)
-				reqBody, err := json.Marshal(externaldataRequest)
-				if err != nil {
-					return externaldata.HandleError(http.StatusInternalServerError, err)
-				}
-
-				req, err := http.NewRequest("POST", provider.Spec.URL, bytes.NewBuffer(reqBody))
-				if err != nil {
-					return externaldata.HandleError(http.StatusInternalServerError, err)
-				}
-				req.Header.Set("Content-Type", "application/json")
-
-				ctx, cancel := context.WithDeadline(bctx.Context, time.Now().Add(time.Duration(provider.Spec.Timeout)*time.Second))
-				defer cancel()
-
-				resp, err := http.DefaultClient.Do(req.WithContext(ctx))
-				if err != nil {
-					return externaldata.HandleError(http.StatusInternalServerError, err)
-				}
-
-				defer resp.Body.Close()
-				respBody, err := ioutil.ReadAll(resp.Body)
-				if err != nil {
-					return externaldata.HandleError(http.StatusInternalServerError, err)
-				}
-
-				var externaldataResponse externaldata.ProviderResponse
-				if err := json.Unmarshal(respBody, &externaldataResponse); err != nil {
-					return externaldata.HandleError(http.StatusInternalServerError, err)
-				}
-
-				regoResponse := externaldata.NewRegoResponse(resp.StatusCode, &externaldataResponse)
-				return externaldata.PrepareRegoResponse(regoResponse)
-			},
-		)
-	}
-	return nil
 }
 
 func copyModules(modules map[string]*ast.Module) map[string]*ast.Module {
